ref: relocate useLinkEditor
